pyproject.toml
[project]
name = "mm-api-integration-guide"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Rishabh Jain", email = "rishabjn10@gmail.com"},
]
dependencies = [
    "certifi==2024.07.04",
    "charset-normalizer==2.1.1",
    "idna==3.4",
    "Pysher==1.0.8",
    "requests==2.28.1",
    "schedule==1.1.0",
    "urllib3==1.26.13",
    "websocket-client==1.4.2",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.156.0",
    "pandas>=2.2.3",
    "data-wrangler>=0.0.3",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
